Update and rename misc.py to code/misc.py
